Fixed clang warnings (build break on some CI)

+++ testing/srt-test-live.cpp
@@ -300,7 +300,7 @@ extern "C" int SrtCheckGroupHook(void* , SRTSOCKET acpsock, int , const sockaddr
         size = sizeof gt;
         if (SRT_ERROR != srt_getsockflag(acpsock, SRTO_GROUPTYPE, &gt, &size))
         {
-            if (gt < Size(gtypes))
+            if (size_t(gt) < Size(gtypes))
                 Verb() << " type=" << gtypes[gt] << VerbNoEOL;
             else
                 Verb() << " type=" << int(gt) << VerbNoEOL;
